% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patamares.r
\name{classfiltrapats}
\alias{classfiltrapats}
\title{Avaliação de patamares}
\usage{
classfiltrapats(dat, tol = c(3, 2, 1.15), plot.dir)
}
\arguments{
\item{dat}{objeto do tipo \code{datpoli} para classificar}

\item{tol}{vetor de tolerancias para o filro sucessivo. Ver Detalhes}

\item{plot.dir}{opcional. Se fornecido, diretorio onde salvar plots dos filtros aplicados a cada
patamar -- pode ser fornecido \code{"auto"}; neste caso a raiz do arquivo de onde foram lidos
os historicos e dados de extensao}
}
\value{
objeto \code{datpoli} contendo item \code{infopats} preenchido e item \code{hist_est}
filtrado de acordo
}
\description{
Classifica o dado em patamares do nível de montante do aproveitamento a jusanta e os filtra
}
\details{
\\TODO: explicar como o processo de filtro funciona
}
\examples{

# filtra vazoes estaveis
dat <- filtravazest(dummydata)

# classifica e filtra patamares
dat <- classfiltrapats(dat)

}
