% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remanso.r
\name{evalremanso}
\alias{evalremanso}
\title{Avaliacao do efeito de remanso}
\usage{
evalremanso(dat, tol = 0.05, step = 2)
}
\arguments{
\item{dat}{objeto do tipo \code{datpoli} ja classificado no qual avaliar remanso}

\item{tol}{tolerancia minima para consideracao de convergencia entre dois patamares. Ver Detalhes}

\item{step}{separacao entre patamares para testar covergencia. Ver Detalhes}
}
\value{
objeto \code{datpoli} contendo item \code{infopats} preenchido com vazoes de convergencia
}
\description{
Identifica quais patamares sofrem influencia parcial ou completa do reservatorio a jusante
}
\details{
//TODO explicar melhor o processo de avaliacao de remanso
}
