% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvabase.r
\name{extraibase}
\alias{extraibase}
\title{Extrator de curva abase}
\usage{
extraibase(dat, subset = NULL)
}
\arguments{
\item{dat}{objeto \code{datpoli} do qual extrair dados para base}

\item{subset}{opcional -- vetor de inteiros ou lógico indicando as posições do dado estável para
usar na curva base. Ver Detalhes}
}
\value{
objeto \code{datcurvabase} -- lista de dois elementos:
\itemize{
\item no primeiro está a parte do dado histórico utilizada para curva base
\item no segundo uma lista de alternativas de dados para extensão do ajuste
}
}
\description{
Facilita a extração dos pontos relativos à curva base a partir do histórico estável
}
\details{
Por padrão a curva base é composta pelos dados sem influência de remanso como identificado
anteriormente em \code{\link{evalremanso}}. Em determinados casos de dados menos consistentes
este processo de identificação não resulta em um conjunto representativo para a curva base, sendo
necessário intervir manualmente.

Para este propósito existe o parâmetro \code{subset}. Se for fornecido, a identificação padrão de
dados da curva base é completamente ignorada e são utilizados apenas os registros informados
através de \code{subset}.
}
