% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patamares.r
\name{extraipats}
\alias{extraipats}
\title{Extração de patamares}
\usage{
extraipats(dat, polibase, min_reg, quais)
}
\arguments{
\item{dat}{objeto do tipo \code{datpoli}}

\item{polibase}{objeto do tipo \code{polijusU} contendo ajuste da curva base}

\item{min_reg}{número mínimo de registros para eligibilidade a ajuste individual}

\item{quais}{vetor de strings indicando quais patamares extrair}
}
\value{
lista de até quatro elementos contendo dados isolados de patamares
}
\description{
Função para extração de até quatro patamares para ajuste individual
}
\details{
A determinação de quais patamares sao ajustados através de curvas individuais se da visando
encontrar um conjunto de curvas que, junto da curva base, estão o mais igualmente distribuídas
pela dispersão de dados observados quanto possível. Atualmente o modelo DECOMP é capaz de lidar
com até cinco famílias de polinômios -- contando a curva base, restam então no máximo quatro
curvas associadas a patamares individuais do reservatório a jusante.
}
