% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvabase.r
\name{extraplog}
\alias{extraplog}
\title{Extrapolação logarítmica}
\usage{
extraplog(dbase, tol = 1e-05)
}
\arguments{
\item{dbase}{objeto \code{datcbase} contendo dados da curva base para extrapolar}

\item{tol}{tolerância para consideração de região linear. Ver Detalhes}
}
\description{
Extrapolação logarítmica
}
