% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vazoesestaveis.r
\name{filtravazest}
\alias{filtravazest}
\title{Filtro de vazões estáveis}
\usage{
filtravazest(dat, n = 6, tol = 0.05)
}
\arguments{
\item{dat}{objeto \code{datpoli}. Ver \code{\link{new_datpoli}}}

\item{n}{inteiro indicando número de registros consecutivos buscados para sequência de vazões
estáveis. Ver Detalhes}

\item{tol}{percentual no formato decimal (e.g. 0.1 para 10\%) de variação máxima tolerada na
sequência para determinação de estabilidade. Ver Detalhes}
}
\value{
objeto \code{datpoli} com item \code{hist_est} preenchido com histórico equivalente
estável. Ver Detalhes
}
\description{
Função que realiza a filtragem de vazões estáveis no dado bruto
}
\details{
A identificação de vazões estáveis segue o processo descrito na
\code{vignette("polijus", package = "polijus")}. De forma geral, são buscados \code{n} registros
horários consecutivos de vazão nos quais a máxima diferença percentual entre o primeiro e os
\code{n - 1} restantes seja inferior ao valor \code{tol}.

A comparação é feita sempre em respeito ao primeiro registro e não de forma sequencial
intencionalmente. Isto busca evitar a consideração incorreta de tomadas de carga lentas (ao longo
de múltiplas horas) como um registro estável. Por exemplo, se \code{n = 6} e \code{tol = .05}

\tabular{rr}{
\eqn{Vazao_1} \tab 10.0\cr
\eqn{Vazao_2} \tab 10.3\cr
\eqn{Vazao_3} \tab 9.80\cr
\eqn{Vazao_4} \tab 9.90\cr
\eqn{Vazao_5} \tab 10.2\cr
\eqn{Vazao_6} \tab 10.1
}

Todos os registros de 2 a 6 variam menos de 5\% em relação ao primeiro, determinando assim uma
sequência de vazões estáveis. Caso estivéssmos numa condição de tomada de carga, com os mesmos
parâmetros de execução

\tabular{rr}{
\eqn{Vazao_1} \tab 10.0\cr
\eqn{Vazao_2} \tab 10.3\cr
\eqn{Vazao_3} \tab 10.7\cr
\eqn{Vazao_4} \tab 11.1\cr
\eqn{Vazao_5} \tab 11.6\cr
\eqn{Vazao_6} \tab 12.0
}

Cada vazão está dentro da faixa de 5\% daquela imediatamente anterior, porém, em função da subida
contínua, o último registro está a 20\% de distância do primeiro.

Uma vez determinadas todas as sequências de \code{n} vazões estáveis no dado, é obtido um
histórico equivalente estável composto das médias de cada \code{n} vazões, níveis de jusante e
níveis de montante do reservatório a jusante.
}
