% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizacao.r
\name{plot.datpoli}
\alias{plot.datpoli}
\title{Plot de objetos \code{datpoli}}
\usage{
\method{plot}{datpoli}(x, qual, legenda = TRUE, ...)
}
\arguments{
\item{x}{objeto \code{datpoli}}

\item{qual}{string indicando o que deve ser plotado. Ver Detalhes}

\item{...}{não é utilizado, existe apenas para compatibilização com a genérica \code{plot}}
}
\value{
plot das informações requisitadas em \code{qual}
}
\description{
Metodo S3 para facilitação dos plots de objetos \code{datpoli}
}
\details{
Exite uma ampla gama de dados contidos em cada objeto \code{datpoli}, de modo que um único
gráfico não é suficiente. Através do parâmetro \code{qual} é possível especificar qual parte do
dado deve ser plotada.

Inicialmente, se são desejados plots do dado como um todo, antes de sua classificação por
patamares, é possível fornecer uma das seguintes opções.

\describe{
\item{\code{qual} = "todos"}{plota dados brutos, estáveis e filtrados}
\item{\code{qual} = "brutos"}{plota apenas dados brutos}
\item{\code{qual} = "estaveis"}{plota apenas dados estáveis}
\item{\code{qual} = "filtrados"}{plota apenas dados filtrados}
}

Deve ser notado que a opção \code{"filtrados"} só é viável após a classificação e filtragem de
dados por patamar através de \code{\link{classfiltrapats}}. Similarmente, a opção \code{"todos"}
só plota os dados filtrados após execução desta função.

A segunda alternativa é o plot de dados por patamar. Neste caso o parâmetro \code{qual} deve ser
fornecido no formato \code{"pat_XXX.X"}. Nesta situação serão plotados apenas os dados referentes
ao patamar \code{XXX.X} colorido para indicar aqueles que foram removido ao longo da filtragem.
}
\examples{

plot(dummydata, "todos") # dado completo
plot(dummydata, "estaveis") # apenas aqueles em condicao estavel
plot(dummydata, "pat_020.5") # plot de um patamar especifico

}
