% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe_polijusU.r
\name{polijusU}
\alias{polijusU}
\alias{new_polijusU}
\alias{print.polijusU}
\alias{coef.polijusU}
\alias{fitted.polijusU}
\alias{predict.polijusU}
\title{Objeto \code{polijusU}}
\usage{
new_polijusU(coefs, bounds, dat, vcov, tipo, tag)

\method{print}{polijusU}(x, ...)

\method{coef}{polijusU}(object, ...)

\method{fitted}{polijusU}(object, derivada = FALSE, ...)

\method{predict}{polijusU}(object, newdata, derivada = FALSE, ...)
}
\arguments{
\item{coefs}{lista de um, dois ou três elementos contendo coeficientes ajustados das partes
polinomiais}

\item{bounds}{lista contendo vetores de duas posições indicando o domínio de cada parte}

\item{dat}{objeto passado à \code{fitpoli}}

\item{vcov}{matriz de variâncias e covariâncas dos coeficientes ajustados}

\item{tipo}{o tipo de ajuste}

\item{tag}{uma tag para identificação de qual curva se trata no print}

\item{...}{demais parâmetros}

\item{object, x}{objeto tipo \code{polijusU}}

\item{derivada}{booleano indicando se deve ser calculada a derivada no lugar do valor ajustado}

\item{newdata}{data.frame ou data.table opcional contendo dados com os quais realizar previsão}
}
\value{
objeto \code{polijusU}, uma lista contendo
\describe{
\item{coef}{lista de coeficientes por parte polinomial ajustada}
\item{bounds}{lista de limites do domínio de cada parte polinomial ajustada}
\item{model}{dado de vazao e nível ajustado com coluna de indicador da parte polinomial}
\item{vcov}{matriz de variâncias e covariâncias dos coeficientes ajustados}
\item{fitted}{vetor de pontos ajustados}
}

Adicionalmente possui atributos
\describe{
\item{npoli}{número de partes polinomiais ajustadas}
\item{graus}{vetor de graus das partes polinomiais}
\item{tipo}{o tipo de ajuste realizado}
\item{tag}{string indicando o nome do ajuste (curva base, patamar XXX, etc)}
}
}
\description{
Contrutor e métodos da classe \code{polijusU} -- curva de jusante univariada
}
\details{
\code{new_polijusU} não deveria ser utilizada diretamente pelo usuário
}
