% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{predictCpp2}
\alias{predictCpp2}
\title{Interpolador/extrpolador linear}
\usage{
predictCpp2(xData, yData, xPred, nExtrap = 2L)
}
\arguments{
\item{xData}{absissas dos dados}

\item{yData}{ordenadas dos dados}

\item{xPred}{absissas para interpolar/extrapolar}

\item{nExtrap}{número de observações utilizadas para extrapolação}
}
\value{
vetor numérico de dados interpolados/extrapolados
}
\description{
Função para interpolar ou extrapolar linearmente em um conjunto de dados
}
