% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leitura.r, R/visualizacao.r
\name{importadados}
\alias{importadados}
\alias{new_datpoli}
\alias{plot.datpoli}
\title{Leitura da planilha padrão}
\usage{
importadados(path = NULL)

new_datpoli(path, cod, hist, ext)

\method{plot}{datpoli}(dat, qual, ...)
}
\arguments{
\item{path}{o caminho da planilha de entrada. Caso não seja fornecido, uma janela do explorador
de arquivos será aberta para busca e seleção}

\item{cod}{escalar inteiro indicando o código da usina}

\item{hist}{dado histórico horário da usina. Ver Detalhes}

\item{ext}{lista de funções para extensão do dado histórico. Ver Detalhes}

\item{dat}{objeto \code{datpoli}}

\item{qual}{string indicando o que deve ser plotado. Ver Detalhes}
}
\value{
objeto \code{datpoli} contendo dados historicos e funções para extensão dos mesmos

objeto \code{datpoli} contendo dados historicos e funções para extensão dos mesmos

plot das informações requisitadas em \code{qual}
}
\description{
Função para leitura e preparo dos dados a partir da planilha front

Função para criação de novas instâncias de objetos da classe datpoli

Metodo S3 para facilitação dos plots de objetos \code{datpoli}
}
\details{
Esta função é uma facilitadora para leitura dos dados de vazão e nível de jusante a partir da
planilha padrão para entrada de dados. Apenas uma facilitadora pois é possível alcançar o mesmo
objetivo lendo os dados necessários a partir de qualquer outra fonte e local -- ver
\code{\link{new_datpoli}} para mais informações.

Tipicamente esta função só é chamada ao final de \code{\link{importadados}} para consolidar as
informações lidas na planilha padrão em um único objeto de saída. Porém, como mencionado na
documentação da mesma, tais objetos não precisam ser criados necessariamente fazendo uso da
planilha padrão, sendo possível lê-los de outros locais e combiná-los com esta função. Para que
isto ocorra sem erros, é necessário garantir algumas características dos dados.

Quanto aos dados históricos, estes devem ser um objeto em formato tabelado (data.frame,
data.table, tibble etc) contendo colunas
\describe{
\item{data}{data da observação em formato \code{Date}}
\item{hora}{hora da observação como inteiro de 1 a 24}
\item{njus}{valor do nível de jusante na data e hora especificada}
\item{vazao}{vazão defluente total (turbinamento + vertimento) da usina}
\item{nmont}{nível de montante do reservatório a jusante}
\item{valido}{booleano indicando a usabilidade do registro horario}
}

Em segundo lugar, o parâmetro \code{ext} deve ser fornecido como uma lista de objetos tabulados
contendo as colunas
\describe{
\item{vazao}{vazão defluente total}
\item{njus}{nível de jusante}
}

Em ambos os casos as colunas podem estar fora de ordem, contanto que todas existam com os nomes
corretos.

Exite uma ampla gama de dados contidos em cada objeto \code{datpoli}, de modo que um único
gráfico não é suficiente. Através do parâmetro \code{qual} é possível especificar qual parte do
dado deve ser plotada.

Inicialmente, se são desejados plots do dado como um todo, antes de sua classificação por
patamares, é possível fornecer uma das seguintes opções.

\describe{
\item{\code{qual} = "todos"}{plota dados brutos, estáveis e filtrados}
\item{\code{qual} = "brutos"}{plota apenas dados brutos}
\item{\code{qual} = "estaveis"}{plota apenas dados estáveis}
\item{\code{qual} = "filtrados"}{plota apenas dados filtrados}
}

Deve ser notado que a opção \code{"filtrados"} só é viável após a classificação e filtragem de
dados por patamar através de \code{\link{classfiltrapats}}. Similarmente, a opção \code{"todos"}
só plota os dados filtrados após execução desta função.

A segunda alternativa é o plot de dados por patamar. Neste caso o parâmetro \code{qual} deve ser
fornecido no formato \code{"pat_XXX.X"}. Nesta situação serão plotados apenas os dados referentes
ao patamar \code{XXX.X} colorido para indicar aqueles que foram removido ao longo da filtragem.
}
\examples{

# Fornecendo um caminho diretamente a funcao
dummypath <- system.file("extdata", "dummydata.xlsx", package = "polijus")
dat <- importadados(path = dummypath)

\dontrun{
# Chamando a funcao sem parametro path para abrir o explorador de arquivos
dat <- importadados()
}


# Dado historico aleatorio
d_hist <- data.frame(vazao = rnorm(240), nmont = rep(seq(10.1, 12, by = .1), each = 12),
    njus = rnorm(240), hora = 1:24, data = as.Date("2020-01-01") + rep(1:10, each = 24),
    valido = TRUE)

# Extensao
l_ext <- list(data.frame(vazao = 1:6, njus = log(1:6)), data.frame(vazao = 0:7, njus = .8 * (0:7) + 1))

# Usando codigo 6 (Furnas) e path NULL (idealmente 'path' aqui seria o caminho onde estao os dados)
new_datpoli(path = NULL, cod = 6, hist = d_hist, ext = l_ext)


# detecta vazoes estaveis
dat <- filtravazest(dummydata)

# classifica e filtra por patamares
dat <- classfiltrapats(dat)

# plot
plot(dat, "todos") # dado completo
plot(dat, "estaveis") # apenas aqueles em condicao estavel
plot(dat, "pat_020.5") # plot de um patamar especifico

\dontrun{

# tentar plots de informacoes ainda nao geradas resulta em erro
plot(dummydata, "filtrados") # dummydata so contem dados brutos, nao existe ainda o historico filtrado
}

}
\seealso{
\code{\link{new_datpoli}} para construtor do objeto de saída

\code{\link{importadados}} para leitura das informações a partir da planilha padrão.
}
